requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.17.0
colorlog==6.7.0
discord.py==2.1.0
python-dotenv==0.21.0
tl3api==1.2.0
